fix user rating
