+++ package.json
@@ -6,7 +6,7 @@
   "dependencies": {
     "@material-ui/core": "^4.9.1",
     "@material-ui/icons": "^4.9.1",
-    "@testrtc/watchrtc-sdk": "@testrtc/watchrtc-sdk@1.34.1-beta.9",
+    "@testrtc/watchrtc-sdk": "1.34.1-beta.9",
     "@types/d3-timer": "^1.0.9",
     "@types/fscreen": "^1.0.1",
     "@types/jest": "^24.9.1",
