+++ src/components/Buttons/EndCallButton/EndCallButton.tsx
@@ -1,13 +1,8 @@
 import React from 'react';
 import clsx from 'clsx';
 import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
-import * as queryString from 'query-string';
-
 import { Button } from '@material-ui/core';
-
 import useVideoContext from '../../../hooks/useVideoContext/useVideoContext';
-
-import watchRTC from '@testrtc/watchrtc-sdk';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -29,20 +24,6 @@ export default function EndCallButton(props: { className?: string }) {
     <Button
       onClick={() => {
         room.disconnect();
-
-        let rating = Math.floor(Math.random() * 5) as any;
-        let message = `User rating is ${rating}`;
-
-        const ratingFromQuery = queryString.parse(window.location.search)?.rating;
-        const ratingMessageFromQuery = queryString.parse(window.location.search)?.ratingMessage;
-
-        if (typeof ratingFromQuery === 'string') {
-          rating = Number(ratingFromQuery);
-        }
-        if (typeof ratingMessageFromQuery === 'string') {
-          message = decodeURI(ratingMessageFromQuery);
-        }
-        watchRTC.setUserRating(rating, message);
       }}
       className={clsx(classes.button, props.className)}
       data-cy-disconnect
